+++ lib/player.rb
@@ -1,7 +1,7 @@
 class Player
   attr_reader :name
 
-  def initialize(name)
+  def initialize(name = 'Player')
     @name = name
   end
 end
