+++ lib/tools.rb
@@ -1,0 +1,7 @@
+def updated_board(board_array)
+  puts "#{board_array[0]} | #{board_array[1]} | #{board_array[2]}"
+  puts '-- --- --'
+  puts "#{board_array[3]} | #{board_array[4]} | #{board_array[5]}"
+  puts '-- --- --'
+  puts "#{board_array[6]} | #{board_array[7]} | #{board_array[8]}"
+end
